+++ src/V1/errors/PackageErrorTable.ts
@@ -31,22 +31,33 @@
 // ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 // POSSIBILITY OF SUCH DAMAGE.
 //
-import { InvalidUuidError, isInvalidUuidError } from "./InvalidUuid";
+import {
+    ErrorTable,
+    ErrorTableTemplateWithNoExtraData,
+    ExtraDataTemplate,
+    NoExtraDataTemplate,
+} from "@ganbarodigital/ts-lib-error-reporting/lib/v1";
+import { httpStatusCodeFrom } from "@ganbarodigital/ts-lib-http-types/lib/v1";
+import { packageNameFrom } from "@ganbarodigital/ts-lib-packagename/lib/v1";
 
+import { InvalidUuidTemplate } from "./InvalidUuid";
 
-describe("isInvalidUuidError", function () {
-    it("accepts InvalidUuidError objects", function() {
-        let inputValue = new InvalidUuidError("12345")
-        expect(isInvalidUuidError(inputValue)).toBeTrue()
-    });
+const PACKAGE_NAME = packageNameFrom("@ganbarodigital/ts-uuid-parser/lib/v1");
 
-    it("rejects other objects", function() {
-        let inputValue = {};
-        expect(isInvalidUuidError(inputValue)).toBeFalse();
-    });
+export class PackageErrorTable implements ErrorTable {
+    [key: string]: ErrorTableTemplateWithNoExtraData<ErrorTable, string, ExtraDataTemplate | NoExtraDataTemplate>;
 
-    it("rejects strings", function() {
-        let inputValue = "";
-        expect(isInvalidUuidError(inputValue)).toBeFalse();
-    });
-});
+    public "invalid-uuid": InvalidUuidTemplate = {
+        packageName: PACKAGE_NAME,
+        errorName: "invalid-uuid",
+        detail: "UUID is invalid / not in RFC 4122 format",
+        status: httpStatusCodeFrom(422),
+        extra: {
+            public: {
+                invalidInput: "the invalid input goes here",
+            },
+        },
+    };
+}
+
+export const ERROR_TABLE = new PackageErrorTable();