+++ src/V1/errors/index.ts
@@ -31,4 +31,5 @@
 // ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 // POSSIBILITY OF SUCH DAMAGE.
 //
+export * from "./PackageErrorTable";
 export * from "./InvalidUuid";