+++ src/V1/guarantees/mustBeUuidData.ts
@@ -30,20 +30,17 @@
 // ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 // POSSIBILITY OF SUCH DAMAGE.
 //
-import { OnError } from "@ganbarodigital/ts-on-error/V1";
+import { OnError, THROW_THE_ERROR } from "@ganbarodigital/ts-lib-error-reporting/lib/v1";
 
-import { InvalidUuidError, invalidUuidError, throwInvalidUuidError } from "../errors/InvalidUuid";
+import { InvalidUuidError } from "../errors/InvalidUuid";
 import { isUuidData } from "../guards/isUuidData";
 
 /**
  * throws an error if the given string is not a well-formatted UUID
  */
-export function mustBeUuidData(input: string, onError?: OnError<InvalidUuidError>): void {
-    // make sure we have an error handler
-    onError = onError ?? throwInvalidUuidError;
-
+export function mustBeUuidData(input: string, onError: OnError = THROW_THE_ERROR): void {
     // a string must contain a well-formatted UUID
     if (!isUuidData(input)) {
-        onError(invalidUuidError, "UUID is invalid / not in RFC 4122 format", new InvalidUuidError(input));
+        onError(new InvalidUuidError({public: { invalidInput: input}}));
     }
 }