+++ src/V1/transforms/unformatted.spec.ts
@@ -31,9 +31,8 @@
 // ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 // POSSIBILITY OF SUCH DAMAGE.
 //
-import { OnError } from "@ganbarodigital/ts-on-error/lib/V1";
+import { OnError } from "@ganbarodigital/ts-lib-error-reporting/lib/v1";
 
-import { InvalidUuidError } from "../errors";
 import { uuidFromFormatted } from "./formatted";
 import { uuidFromUnformatted, uuidToUnformatted } from "./unformatted";
 
@@ -60,13 +59,13 @@ describe("uuidFromUnformatted()", () => {
     });
 
     it("rejects a string that's shorter than 32 characters", () => {
-        const expectedDescription = "input string is wrong length; must be 32";
+        const expectedDescription = "UUID is invalid / not in RFC 4122 format";
         let actualDescription = "";
 
-        const onError: OnError<InvalidUuidError> = (reason, description, extra) => {
-            actualDescription = description;
-            throw extra;
-        }
+        const onError: OnError = (e) => {
+            actualDescription = e.details.detail;
+            throw e;
+        };
         const inputValue = "this is a short string";
 
         expect(() => {uuidFromUnformatted(inputValue, onError); } ).toThrow();
@@ -74,13 +73,13 @@ describe("uuidFromUnformatted()", () => {
     });
 
     it("rejects a string that's longer than 32 characters", () => {
-        const expectedDescription = "input string is wrong length; must be 32";
+        const expectedDescription = "UUID is invalid / not in RFC 4122 format";
         let actualDescription = "";
 
-        const onError: OnError<InvalidUuidError> = (reason, description, extra) => {
-            actualDescription = description;
-            throw extra;
-        }
+        const onError: OnError = (e) => {
+            actualDescription = e.details.detail;
+            throw e;
+        };
         const inputValue = "this is a long string that's definitely longer than a UUID is";
 
         expect(() => {uuidFromUnformatted(inputValue, onError); } ).toThrow();